Redesigned filter buttons to use less space, and put them inline with the table toolbars when possible
Part of this involved creating a way to extend the toolbars on tables.  That design has been used in a few other places as well.

A TODO is to get the filter bars to center when there's space, but then be pushed over when there isn't.  David Mohs suggested looking at flex to do this.