Refresh current tab by clicking it
Used :component-will-receive-props to accomplish this, however some tabs were also triggering
reloads on :component-did-update, which was causing multiple loads.  So I added a simple locking
mechanism to each tab to prevent simultaneous data load calls.  This is a bit of a hack; a future
improvement would be to modify the TabBar component to trigger a specific reload rather than
relying on React lifecycle methods.

Also fixed an issue with the tab bar's margins, and standardized some padding and the centering
of the "loading..." spinner across all tabs.